% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_file_ext}
\alias{check_file_ext}
\title{Check if a file has a specific extension (case-insensitive)}
\usage{
check_file_ext(file, extension)
}
\arguments{
\item{file}{A character string specifying the path to the file.}

\item{extension}{A character string specifying the expected file extension (without the dot).}
}
\value{
A logical value: TRUE if the file has the specified extension, FALSE otherwise.
}
\description{
This function checks if a given file has a specific extension, ignoring case.
}
