% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collect_data_list_paths}
\alias{collect_data_list_paths}
\title{Collect file paths based on file names without extensions}
\usage{
collect_data_list_paths(
  sub_dir,
  file_names,
  use_wd,
  prefer_sas,
  prefer_rds = FALSE
)
}
\arguments{
\item{sub_dir}{A relative directory/folder that will be appended to a base path defined by \code{Sys.getenv("RXD_DATA")}.
If the argument is left as NULL, the function will load data from the working directory \code{getwd()}.}

\item{file_names}{CDISC names for the files}

\item{use_wd}{for "use working directory" - a flag used when importing local files
not on NFS - default value is FALSE}

\item{prefer_sas}{if TRUE, imports \code{.sas7bdat} files first instead of
\code{.parquet} and \code{.rds} files}

\item{prefer_rds}{if TRUE, imports \code{.rds} files first instead of
\code{.parquet} and \code{.sas7bdat} files}
}
\value{
a character vector of resolved file paths
}
\description{
Constructs a list of file paths based on an input vector of file names without extensions.
Preference is given to \code{.parquet} files, if present, over \code{.rds} and
\code{.sas7bdat} files unless \code{prefer_sas} or \code{prefer_rds} is \code{TRUE}.
\code{prefer_sas} and \code{prefer_rds} are mutually exclusive.
}
