% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_info}
\alias{file_info}
\title{Extract file information based on file.info()}
\usage{
file_info(file)
}
\arguments{
\item{file}{A character string specifying the path to the file.}
}
\value{
A list containing file information from file.info(file, extra_cols = FALSE) and the path and file name.
}
\description{
This function extracts file information from a given file.
}
