% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvloader.R
\name{get_base_dir}
\alias{get_base_dir}
\title{Get Base Directory Path}
\usage{
get_base_dir(env_var)
}
\arguments{
\item{env_var}{[character(1)] The name of the environment variable containing the base directory path.}
}
\value{
[character(1)] The normalized path to the base directory.
}
\description{
This function retrieves the base directory path from a specified environment variable.
It checks if the environment variable is set and if the directory exists.
}
\examples{
# Create a temporary directory
temp_dir <- tempdir()

# Set the BASE_DIR environment variable
Sys.setenv(BASE_DIR = temp_dir)

# Get the base directory path
dv.loader:::get_base_dir("BASE_DIR")

}
\keyword{internal}
