% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvloader.R
\name{get_cre_path}
\alias{get_cre_path}
\title{Get CRE Path}
\usage{
get_cre_path()
}
\value{
[character(1)] The path to the CRE directory.
}
\description{
This function retrieves the path to the CRE (Clinical Research Environment) directory.
It uses the "RXD_DATA" environment variable as the base directory.
}
