% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_paths}
\alias{get_file_paths}
\title{Get File Paths}
\usage{
get_file_paths(dir_path, file_names, prefer_sas = FALSE)
}
\arguments{
\item{dir_path}{[character(1)] The directory path where the files are located.}

\item{file_names}{[character(1+)] A vector of file names to process.}

\item{prefer_sas}{[logical(1)] Whether to prefer SAS files over RDS files. Default is FALSE.}
}
\value{
[character] A vector of normalized file paths.
}
\description{
This function constructs file paths for given file names, handling both RDS and SAS7BDAT files.
It can prioritize SAS files over RDS files based on the `prefer_sas` parameter.
}
\examples{
temp_dir <- tempdir()

file_names <- c("adsl", "adae")

file.create(file.path(temp_dir, paste0(file_names, ".rds")))
file.create(file.path(temp_dir, paste0(file_names, ".sas7bdat")))

list.files(temp_dir)

get_file_paths(dir_path = temp_dir, file_names = file_names)
get_file_paths(dir_path = temp_dir, file_names = file_names, prefer_sas = TRUE)

unlink(temp_dir, recursive = TRUE)

}
