% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvloader.R
\name{get_nfs_path}
\alias{get_nfs_path}
\title{Get NFS Path}
\usage{
get_nfs_path(env_var = "RXD_DATA")
}
\arguments{
\item{env_var}{[character(1)] The environment variable name for the base directory. Default is "RXD_DATA".}
}
\value{
[character(1)] The path to the NFS directory.
}
\description{
This function retrieves the path to the NFS (Network File System) directory.
}
