% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data_files}
\alias{load_data_files}
\title{Load Data Files}
\usage{
load_data_files(file_paths)
}
\arguments{
\item{file_paths}{[character(1+)] A vector of file paths to read.}
}
\value{
[list] A named list of data frames, where each name is the basename of the corresponding file path.
}
\description{
This function reads data from multiple file paths and returns a list of data frames.
It supports reading RDS and SAS7BDAT files.
}
\examples{
path <- system.file("examples", "iris.sas7bdat", package = "haven")
data_list <- load_data_files(file_paths = path)
str(data_list)

}
