% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rds.R
\name{load_rds}
\alias{load_rds}
\title{Load RDS files}
\usage{
load_rds(files)
}
\arguments{
\item{files}{A character vector of file paths to RDS files.}
}
\value{
A list of data frames, each containing the data from an RDS file.
}
\description{
This function loads RDS files via readRDS() and returns a list of data frames.
}
\examples{
# Create temporary directory and files
temp_dir <- tempdir()
adsl_rds_file <- file.path(temp_dir, "adsl.rds")
adae_rds_file <- file.path(temp_dir, "adae.rds")

# Write example data to RDS files
saveRDS(pharmaverseadam::adsl, adsl_rds_file)
saveRDS(pharmaverseadam::adae, adae_rds_file)

# Load RDS files
rds_data_list <- load_rds(c(adsl_rds_file, adae_rds_file))

# Clean up
unlink(c(adsl_rds_file, adae_rds_file))
}
