% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sas.R
\name{load_sas}
\alias{load_sas}
\title{Load SAS files}
\usage{
load_sas(files)
}
\arguments{
\item{files}{A character vector of file paths to SAS files.}
}
\value{
A list of data frames, each containing the data from a SAS file.
}
\description{
This function loads SAS files via haven::read_sas() and returns a list of data frames.
}
\examples{
# Create temporary directory and files
temp_dir <- tempdir()
adsl_sas_file <- file.path(temp_dir, "adsl.sas7bdat")
adae_sas_file <- file.path(temp_dir, "adae.sas7bdat")

# Write example data to SAS files
haven::write_sas(pharmaverseadam::adsl, adsl_sas_file)
haven::write_sas(pharmaverseadam::adae, adae_sas_file)

# Load SAS files
sas_data_list <- load_sas(c(adsl_sas_file, adae_sas_file))

# Clean up
unlink(c(adsl_sas_file, adae_sas_file))
}
