% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_xpt.R
\name{load_xpt}
\alias{load_xpt}
\title{Load XPT files}
\usage{
load_xpt(files)
}
\arguments{
\item{files}{A character vector of file paths to XPT files.}
}
\value{
A list of data frames, each containing the data from an XPT file.
}
\description{
This function loads XPT files via haven::read_xpt() and returns a list of data frames.
}
\examples{
# Create temporary directory and files
temp_dir <- tempdir()
adsl_xpt_file <- file.path(temp_dir, "adsl.xpt")
adae_xpt_file <- file.path(temp_dir, "adae.xpt")

# Write example data to XPT files
haven::write_xpt(pharmaverseadam::adsl, adsl_xpt_file)
haven::write_xpt(pharmaverseadam::adae, adae_xpt_file)

# Load XPT files
xpt_data_list <- load_xpt(c(adsl_xpt_file, adae_xpt_file))

# Clean up
unlink(c(adsl_xpt_file, adae_xpt_file))  
}
