% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_file}
\alias{read_file}
\title{Reads RDS/SAS file and metadatas from first 6 items from file.info() its file path}
\usage{
read_file(file_path, file_name)
}
\arguments{
\item{file_path}{a path to a file}

\item{file_name}{name of a file}
}
\value{
a data object with an extra attribute of metadata
}
\description{
Reads RDS/SAS file and metadatas from first 6 items from file.info() its file path
}
