% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvloader.R
\name{reduce_data_frame_memory_footprint}
\alias{reduce_data_frame_memory_footprint}
\title{Transform data.frame columns to use leaner types}
\usage{
reduce_data_frame_memory_footprint(df)
}
\arguments{
\item{df}{\verb{[data.frame]} Data frame to transform}
}
\value{
\verb{[data.frame]} Transformed data frame
}
\description{
Transforms character columns into factors and numeric content into integer vectors, when it does not
lead to loss of precision.
}
