% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collect_data_list_paths}
\alias{collect_data_list_paths}
\title{For each file name provided, reads in the first matching file and its meta data/attributes.
Preference is given to RDS because its faster}
\usage{
collect_data_list_paths(sub_dir, file_names, use_wd, prefer_sas)
}
\arguments{
\item{sub_dir}{A relative directory/folder that will be appended to a base path defined by \code{Sys.getenv("RXD_DATA")}.
If the argument is left as NULL, the function will load data from the working directory \code{getwd()}.}

\item{file_names}{CDISC names for the files}

\item{use_wd}{for "use working directory" - a flag used when importing local files
not on NFS - default value is FALSE}

\item{prefer_sas}{if TRUE, imports .sas7bdat files first instead of .RDS files}
}
\value{
returns a list of dataframes with metadata as an attribute on each dataframe
}
\description{
For each file name provided, reads in the first matching file and its meta data/attributes.
Preference is given to RDS because its faster
}
