% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_data_list}
\alias{create_data_list}
\title{For each file name provided, reads in the first matching file and its meta data/attributes.
Preference is given to RDS because its faster}
\usage{
create_data_list(file_path, file_names, prefer_sas)
}
\arguments{
\item{file_path}{the folder where the files are}

\item{file_names}{CDISC names for the files}

\item{prefer_sas}{if TRUE, imports .sas7bdat files first instead of .RDS files}
}
\value{
returns a list of dataframes with metadata as an attribute on each dataframe
}
\description{
For each file name provided, reads in the first matching file and its meta data/attributes.
Preference is given to RDS because its faster
}
