% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvloader.R
\name{get_cre_path}
\alias{get_cre_path}
\title{Get Base Path from an Environment Variable}
\usage{
get_cre_path()
}
\value{
[character(1)] The normalized base path.
}
\description{
This function is an alias for `get_nfs_path()` to maintain backwards compatibility.
}
