% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_cre_path}
\alias{get_cre_path}
\title{gets the NFS base path from an env var
alias for get_nfs_path to maintain backwards compatibility}
\usage{
get_cre_path()
}
\description{
gets the NFS base path from an env var
alias for get_nfs_path to maintain backwards compatibility
}
