% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvloader.R
\name{get_nfs_path}
\alias{get_nfs_path}
\title{Get Base Path from an Environment Variable}
\usage{
get_nfs_path()
}
\value{
[character(1)] The normalized base path.
}
\description{
This function assumes that there is an environment variable called `RXD_DATA` 
which is set to the base path of the data directory.
}
