% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvloader.R
\name{get_nfs_path}
\alias{get_nfs_path}
\title{gets the NFS base path from an env var
It assumes there is an env var
called RXD_DATA which holds the path suffix.}
\usage{
get_nfs_path()
}
\value{
the NFS base path
}
\description{
gets the NFS base path from an env var
It assumes there is an env var
called RXD_DATA which holds the path suffix.
}
