% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvloader.R
\name{load_data}
\alias{load_data}
\title{Load Data Files}
\usage{
load_data(
  sub_dir = NULL,
  file_names,
  use_wd = FALSE,
  prefer_sas = FALSE,
  env_var = "RXD_DATA",
  print_file_paths = FALSE
)
}
\arguments{
\item{sub_dir}{[character(1)] Optional character string specifying a subdirectory. Default is NULL.}

\item{file_names}{[character(1+)] Character vector of file names to load (without extension).}

\item{use_wd}{[logical(1)] Logical indicating whether to use the current working directory. Default is FALSE.}

\item{prefer_sas}{[logical(1)] Logical indicating whether to prefer SAS7BDAT files over RDS. Default is FALSE.}

\item{env_var}{[character(1)] The environment variable name for the base directory. Default is "RXD_DATA".}

\item{print_file_paths}{[logical(1)] Logical indicating whether to print the directory path and file names. Default is FALSE.}
}
\value{
A named list of data frames, where each name corresponds to a loaded file.
}
\description{
This function loads data files from a specified directory or the current working directory.
It supports loading both RDS and SAS7BDAT files.
}
\examples{
# Get the current value of the RXD_DATA environment variable
base_dir <- Sys.getenv("RXD_DATA")

# Set the RXD_DATA environment variable to the path of the haven package
Sys.setenv(RXD_DATA = find.package("haven"))

data_list <- load_data(sub_dir = "examples", file_names = c("iris.sas7bdat"))
str(data_list)

# Reset the RXD_DATA environment variable to its original value
Sys.setenv(RXD_DATA = base_dir)

}
