% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvloader.R
\name{load_data}
\alias{load_data}
\title{Loads data into memory based on study directory and one or more file_names.}
\usage{
load_data(sub_dir = NULL, file_names, use_wd = FALSE, prefer_sas = FALSE)
}
\arguments{
\item{sub_dir}{A relative directory/folder that will be appended to a base path defined by `Sys.getenv("RXD_DATA")`. 
If the argument is left as NULL (environment variable "RXD_DATA" not pre-defined), the function 
will use the working directory as the `sub_dir`.}

\item{file_names}{Study file or file_names name(s) - can be a vector of strings.
This is the only required argument.}

\item{use_wd}{for "use working directory" - a flag used when importing local files
not on NFS - default value is FALSE}

\item{prefer_sas}{if set to TRUE, imports sas7bdat files first before looking for
RDS files (the opposite of default behavior)}
}
\value{
a list of dataframes
}
\description{
Loads data into memory based on study directory and one or more file_names.
}
\examples{
\dontrun{
test_data_path <- "../inst/extdata/"
data_list <- load_data(
  sub_dir = test_data_path,
  file_names = "dummyads2",
  use_wd = TRUE
)
}
}
