% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvloader.R
\name{load_files}
\alias{load_files}
\title{Load data files from explicit paths}
\usage{
load_files(file_paths, reduce_memory_footprint = TRUE)
}
\arguments{
\item{file_paths}{\verb{[character(1+)]} Files to read. Optionally named.}
}
\value{
\verb{[list]} A named list of data frames, where each name is either:
\itemize{
\item the name associated to the element in the \code{file_paths} argument, or, if not provided...
\item the name of the file itself, after stripping it of its leading path and trailing extension
}
}
\description{
Read data from provided paths and return it as a list of data frames.
Supports both .rds and .sas7bdat formats.
}
