% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvloader.R
\name{load_files}
\alias{load_files}
\title{Load data files from explicit paths}
\usage{
load_files(file_paths)
}
\arguments{
\item{file_paths}{[character(1+)] Files to read. Optionally named.}
}
\value{
[list] A named list of data frames, where each name is either:
 - the name associated to the element in the `file_paths` argument, or, if not provided...
 - the name of the file itself, after stripping it of its leading path and trailing extension
}
\description{
Read data from provided paths and return it as a list of data frames.
Supports both .rds and .sas7bdat formats.
}
