% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvloader.R
\name{memory_use_report}
\alias{memory_use_report}
\title{Print data remapping report of the transformations performed by \code{reduce_data_frame_memory_footprint}}
\usage{
memory_use_report(df)
}
\arguments{
\item{df}{\verb{[data.frame]} Output from \code{reduce_data_frame_memory_footprint}}
}
\value{
\verb{[character(1)]} Report
}
\description{
Print data remapping report of the transformations performed by \code{reduce_data_frame_memory_footprint}
}
