% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvloader.R
\name{reduce_column_memory_footprint}
\alias{reduce_column_memory_footprint}
\title{Transform data.frame column to use leaner types}
\usage{
reduce_column_memory_footprint(col_data)
}
\arguments{
\item{col_data}{Vector to transform}
}
\value{
Transformed vector
}
\description{
Transforms a character column into a factor or a numeric column into integer, when the transformation does not
lead to loss of precision.
}
